requirements.txt
Django==5.1.5
pillow==11.1.0
djangorestframework==3.15.2